GearUp: Integratest latest from dev.
[git-p4: depot-paths = "//depot/traktor/": change = 9068]

+++ code/Online/Steam/SteamStatistics.h
@@ -20,7 +20,7 @@ class SteamStatistics : public IStatisticsProvider
 public:
 	SteamStatistics(SteamSessionManager* sessionManager, const std::list< std::wstring >& statIds);
 
-	virtual bool enumerate(std::map< std::wstring, float >& outStats);
+	virtual bool enumerate(std::map< std::wstring, int32_t >& outStats);
 
 	virtual bool set(const std::wstring& statId, int32_t value);
